requirements.txt
asgiref==3.8.1
aspose-cells==24.5.0
aspose-words==24.5.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
idna==3.7
JPype1==1.5.0
packaging==24.0
panda==0.3.1
plum-dispatch==1.7.4
pypandoc==1.13
requests==2.32.0
Spire.Doc==12.4.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
workbook==1.1
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
